blog/requirements.txt
django
django-rest-framework
Pillow
